small fixes and new _operations, look in operation_dispatcher
Former-commit-id: 85d443df332177a4754e4f28d673d4f48885685d

+++ virtual_machine/src/carriage/_carriage.h
@@ -34,7 +34,8 @@ typedef struct				s_carriage
 	int				is_correct_op_code;
 	int				arena_position;
 //	int				bytes_to_next_op;
-	long int		registers[REG_NUMBER];
+	int	        	registers[REG_NUMBER];
+//	long int		registers[REG_NUMBER];
 }							t_carriage;
 
 void                _carriage_set_op_code(t_carriage *self, t_arena *arena);
